[RF] Implement host and device buffers in dynamically loaded library
Like this, we can avoid a few more preprocessor checks conditional on
`cuda=ON`.
