+++ roofit/batchcompute/res/RooBatchCompute.h
@@ -113,6 +113,28 @@ struct ReduceNLLOutput {
    std::size_t nNaNValues = 0;
 };
 
+class AbsBuffer {
+public:
+   virtual ~AbsBuffer() = default;
+
+   virtual double const *cpuReadPtr() const = 0;
+   virtual double const *gpuReadPtr() const = 0;
+
+   virtual double *cpuWritePtr() = 0;
+   virtual double *gpuWritePtr() = 0;
+};
+
+class AbsBufferManager {
+public:
+   virtual ~AbsBufferManager() = default;
+
+   virtual std::unique_ptr<AbsBuffer> makeScalarBuffer() = 0;
+   virtual std::unique_ptr<AbsBuffer> makeCpuBuffer(std::size_t size) = 0;
+   virtual std::unique_ptr<AbsBuffer> makeGpuBuffer(std::size_t size) = 0;
+   virtual std::unique_ptr<AbsBuffer>
+   makePinnedBuffer(std::size_t size, RooFit::Detail::CudaInterface::CudaStream *stream = nullptr) = 0;
+};
+
 /**
  * \class RooBatchComputeInterface
  * \ingroup roofit_dev_docs_batchcompute
@@ -144,6 +166,8 @@ public:
 
    virtual Architecture architecture() const = 0;
    virtual std::string architectureName() const = 0;
+
+   virtual std::unique_ptr<AbsBufferManager> createBufferManager() const = 0;
 };
 
 /**
@@ -193,35 +217,6 @@ inline ReduceNLLOutput reduceNLL(Config cfg, std::span<const double> probas, std
    auto dispatch = cfg.useCuda() ? dispatchCUDA : dispatchCPU;
    return dispatch->reduceNLL(cfg, probas, weights, offsetProbas);
 }
-
-class AbsBuffer {
-public:
-   virtual ~AbsBuffer() = default;
-
-   virtual double const *cpuReadPtr() const = 0;
-   virtual double const *gpuReadPtr() const = 0;
-
-   virtual double *cpuWritePtr() = 0;
-   virtual double *gpuWritePtr() = 0;
-};
-
-struct BufferQueuesMaps;
-
-class BufferManager {
-
-public:
-   BufferManager();
-   ~BufferManager();
-
-   std::unique_ptr<AbsBuffer> makeScalarBuffer();
-   std::unique_ptr<AbsBuffer> makeCpuBuffer(std::size_t size);
-   std::unique_ptr<AbsBuffer> makeGpuBuffer(std::size_t size);
-   std::unique_ptr<AbsBuffer>
-   makePinnedBuffer(std::size_t size, RooFit::Detail::CudaInterface::CudaStream *stream = nullptr);
-
-private:
-   std::unique_ptr<BufferQueuesMaps> _queuesMaps;
-};
 
 } // End namespace RooBatchCompute
